Create DSImg.ts
