+++ src/lib/dsLib.ts
@@ -13,13 +13,3 @@ export function nvram_get(key: string): string | null {
 export function nvram_clear() {
     localStorage.clear();
 }
-
-export async function load_image(url: string): Promise<HTMLImageElement> {
-    return new Promise((resolve, reject) => {
-        const img = new Image();
-        img.src = url;
-
-        img.onload = () => resolve(img);
-        img.onerror = (error) => reject(error);
-    });
-}
